layout/style/test/test_bug721136.html

<!doctype html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=721136
-->
<title>Test for Bug 721136</title>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=721136">Mozilla Bug 721136</a>
<pre id="test">
<script>
[
  [" mAtRiX(1, 2,3,4, 5,6 ) ", "matrix(1, 2, 3, 4, 5, 6)"],
  [" mAtRiX3d( 1,2,3,0,4 ,5,6,0,7,8 , 9,0,10, 11,12,1 )  ",
   "matrix3d(1, 2, 3, 0, 4, 5, 6, 0, 7, 8, 9, 0, 10, 11, 12, 1)"],
  [" pErSpEcTiVe( 400Px ) ", "perspective(400px)"],
  [" rOtAtE( 90dEg ) ", "rotate(90deg)"],
  [" rOtAtE3d( 0,0 , 1 ,180DeG ) ", "rotate3d(0, 0, 1, 180deg)"],
  [" rOtAtEx( 100GrAD ) ", "rotateX(100grad)"],
  [" rOtAtEy( 1.57RaD ) ", "rotateY(1.57rad)"],
  [" rOtAtEz( 0.25TuRn ) ", "rotateZ(0.25turn)"],
  [" sCaLe( 2 ) ", "scale(2)"],
  [" sCaLe( 2,3 ) ", "scale(2, 3)"],
  [" sCaLe3D( 2,4 ,  -9 )  ", "scale3d(2, 4, -9)"],
  [" sCaLeX( 2 ) ", "scaleX(2)"],
  [" sCaLeY( 2 ) ", "scaleY(2)"],
  [" sCaLeZ( 2 ) ", "scaleZ(2)"],
  [" sKeW( 45dEg ) ", "skew(45deg)"],
  [" sKeW( 45dEg,45DeG ) ", "skew(45deg, 45deg)"],
  [" sKeWx( 45DeG ) ", "skewX(45deg)"],
  [" sKeWy( 45DeG ) ", "skewY(45deg)"],
  [" tRaNsLaTe( 1Px ) ", "translate(1px)"],
  [" tRaNsLaTe( 1Px,3Pt ) ", "translate(1px, 3pt)"],
  [" tRaNsLaTe3D( 21pX,-6pX , 4pX )  ", "translate3d(21px, -6px, 4px)"],
  [" tRaNsLaTeX( 1pT ) ", "translateX(1pt)"],
  [" tRaNsLaTeY( 1iN ) ", "translateY(1in)"],
  [" tRaNsLaTeZ( 15.4pX ) ", "translateZ(15.4px)"],
  ["tranSlatex( 16px )rotatez(-90deg)  rotate(100grad)\ttranslate3d(12pt, 0pc, 0.0em)",
   "translateX(16px) rotateZ(-90deg) rotate(100grad) translate3d(12pt, 0pc, 0em)"],
].forEach(function(arr) {
  document.documentElement.style.MozTransform = arr[0];
  is(document.documentElement.style.MozTransform, arr[1],
    "incorrect serialization");
});
</script>
</pre>
